Added few things:
- Setup ESLint with a popular style guide to enforce code style.
    - Make vscode recogmise this settings and do as is.

- Setup Vue-Router, routed the following routes:
    - `/` -> `src/views/Home.vue`
    - `/login` -> `src/views/Login.vue`
