+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+  "eslint.format.enable": true,
+  "editor.formatOnSave": true,
+  "editor.tabSize": 2
+}
