--- login.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" type="text/css" href="styles.css" />
-
-    <title>Dwello Login</title>
-  </head>
-
-  <body>
-    <div class="main">
-      <div class="left-center">
-        <h1 class="login-logo">Dwello<span class="small-r">®</span></h1>
-      </div>
-      <div class="right-center">
-        <div class="login-rectangle"></div>
-      </div>
-    </div>
-  </body>
-</html>
