+++ src/main.js
@@ -1,5 +1,24 @@
 import { createApp } from "vue";
-import "./style.css";
+import { createRouter, createWebHistory } from "vue-router";
+
 import App from "./App.vue";
 
-createApp(App).mount("#app");
+import Home from "./views/Home.vue";
+import Login from "./views/Login.vue";
+
+import "./style.css";
+
+const routes = [
+  { path: "/", component: Home },
+  { path: "/login", component: Login },
+];
+
+const router = createRouter({
+  history: createWebHistory(),
+  routes,
+});
+
+const app = createApp(App);
+app.use(router);
+
+app.mount("#app");
