+++ src/views/Login.vue
@@ -1,0 +1,10 @@
+<template>
+  <div class="main">
+    <div class="left-center">
+      <h1 class="login-logo">Dwello<span class="small-r">®</span></h1>
+    </div>
+    <div class="right-center">
+      <div class="login-rectangle"></div>
+    </div>
+  </div>
+</template>
